Added slime acid pool mechanic, as well as blobs (particles). Changed health to be a float and damage over time now takes frame time into account (for proper scaling, regardless of framerate). Added a getZIndex() method to sort sprites (floor sprites return -1, instead of 0, so they are drawn first). Added a ton of ugly code on the SlimeAcid class; I need to clean that up, probably splitting it into further classes. I also need to add a basic particle class.

+++ core/src/com/dungeon/engine/render/Tileset.java
@@ -16,12 +16,12 @@ public abstract class Tileset {
 		this.tile_size = tile_size;
 	}
 
-	public TextureRegion getTile(int x, int y) {
-		return new TextureRegion(texture, tile_size * x, tile_size * y, tile_size, tile_size);
+	public TextureRegion getTile(float x, float y) {
+		return new TextureRegion(texture, (int) (tile_size * x), (int) (tile_size * y), tile_size, tile_size);
 	}
 
-	public TextureRegion getTile(int x, int y, int x_tiles, int y_tiles) {
-		return new TextureRegion(texture, tile_size * x, tile_size * y, tile_size * x_tiles, tile_size * y_tiles);
+	public TextureRegion getTile(float x, float y, float x_tiles, float y_tiles) {
+		return new TextureRegion(texture, (int) (tile_size * x), (int) (tile_size * y), (int) (tile_size * x_tiles), (int) (tile_size * y_tiles));
 	}
 
 	protected Tile[] createTiles(TextureRegion ... regions) {
